Move HelpMessageBuilder to new util.messages package
